feat: find/replace takes current filtering into account
