+++ src/hooks/resourcesHooks/useResourceYamlSchema.ts
