+++ src/navsections/ClusterDiffSectionBlueprint/ClusterDiffSectionEmptyDisplay.tsx
@@ -1,17 +1,9 @@
-import React, {useMemo} from 'react';
+import React from 'react';
 
-import {useAppSelector} from '@redux/hooks';
+import useResourceFilter from '@hooks/resourcesHooks/useResoucreFilter';
 
 function ClusterDiffSectionEmptyDisplay() {
-  const resourceFilter = useAppSelector(state => state.main.resourceFilter);
-
-  const appliedFilters = useMemo(() => {
-    return Object.entries(resourceFilter)
-      .map(([key, value]) => {
-        return {filterName: key, filterValue: value};
-      })
-      .filter(filter => filter.filterValue && Object.values(filter.filterValue).length);
-  }, [resourceFilter]);
+  const {appliedFilters} = useResourceFilter();
 
   if (appliedFilters.length === 0) {
     return null;
